requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
boto3==1.21.42
botocore==1.24.42
click==8.1.2
Django==3.2.12
django-storages==1.12.3
djangorestframework==3.13.1
gunicorn==20.1.0
importlib-metadata==4.11.3
jmespath==1.0.0
joblib==1.1.0
nltk==3.7
numpy==1.21.6
pandas==1.3.5
psycopg2-binary==2.9.3
python-dateutil==2.8.2
pytz==2022.1
regex==2022.3.15
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
textblob==0.17.1
tqdm==4.64.0
typing-extensions==4.1.1
urllib3==1.26.9
zipp==3.8.0
